Split grafana client creation in its own folder
